Split BSD-licensed code in own module
This is cleaner and makes tracking licenses much easier.
